+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Remove usage of the `org.realityforge.anodoc` dependency.
+* Remove usage of un-enforced `javax.annotation.Nonnegative` annotation.
 
 ### [v0.17](https://github.com/arez/arez-ticker/tree/v0.17) (2018-06-13)
 [Full Changelog](https://github.com/arez/arez-ticker/compare/v0.16...v0.17)